Move scope-related classes to `evaluation.scope`
