+++ src/main/kotlin/sigma/evaluation/values/Closure.kt
@@ -3,9 +3,9 @@ package sigma.evaluation.values
 import sigma.Thunk
 import sigma.syntax.expressions.ExpressionTerm
 import sigma.syntax.typeExpressions.TupleTypeLiteralTerm
-import sigma.evaluation.values.tables.Scope
+import sigma.evaluation.scope.Scope
 import sigma.evaluation.values.tables.Table
-import sigma.evaluation.values.tables.chainWith
+import sigma.evaluation.scope.chainWith
 
 class Closure(
     private val context: Scope,
