+++ src/main/kotlin/sigma/syntax/expressions/UnorderedTupleLiteralTerm.kt
@@ -8,7 +8,7 @@ import sigma.syntax.SourceLocation
 import sigma.semantics.types.UnorderedTupleType
 import sigma.evaluation.values.PrimitiveValue
 import sigma.evaluation.values.tables.DictTable
-import sigma.evaluation.values.tables.Scope
+import sigma.evaluation.scope.Scope
 import sigma.semantics.types.Type
 import sigma.evaluation.values.Symbol
 import sigma.evaluation.values.TypeErrorException
