+++ src/main/kotlin/sigma/syntax/typeExpressions/OrderedTupleTypeLiteralTerm.kt
@@ -8,7 +8,7 @@ import sigma.parser.antlr.SigmaParser.OrderedTupleTypeLiteralContext
 import sigma.semantics.types.OrderedTupleType
 import sigma.evaluation.values.IntValue
 import sigma.evaluation.values.Symbol
-import sigma.evaluation.values.tables.Scope
+import sigma.evaluation.scope.Scope
 import sigma.evaluation.values.tables.Table
 
 data class OrderedTupleTypeLiteralTerm(
