+++ src/main/kotlin/sigma/syntax/typeExpressions/TupleTypeLiteralTerm.kt
@@ -6,7 +6,7 @@ import sigma.parser.antlr.SigmaParser.OrderedTupleTypeLiteralContext
 import sigma.parser.antlr.SigmaParser.UnorderedTupleTypeLiteralContext
 import sigma.parser.antlr.SigmaParserBaseVisitor
 import sigma.semantics.types.TupleType
-import sigma.evaluation.values.tables.Scope
+import sigma.evaluation.scope.Scope
 import sigma.evaluation.values.tables.Table
 
 abstract class TupleTypeLiteralTerm : TypeExpressionTerm() {
