+++ src/main/kotlin/sigma/syntax/typeExpressions/UnorderedTupleTypeLiteralTerm.kt
@@ -8,7 +8,7 @@ import sigma.semantics.types.UnorderedTupleType
 import sigma.evaluation.values.PrimitiveValue
 import sigma.evaluation.values.Symbol
 import sigma.evaluation.values.TypeErrorException
-import sigma.evaluation.values.tables.Scope
+import sigma.evaluation.scope.Scope
 import sigma.evaluation.values.tables.Table
 
 data class UnorderedTupleTypeLiteralTerm(
