+++ src/test/kotlin/sigma/syntax/expressions/FieldReadTests.kt
@@ -13,7 +13,7 @@ import sigma.evaluation.values.IntValue
 import sigma.evaluation.values.Symbol
 import sigma.evaluation.values.TypeErrorException
 import sigma.evaluation.values.tables.DictTable
-import sigma.evaluation.values.tables.FixedScope
+import sigma.evaluation.scope.FixedScope
 import kotlin.test.Test
 import kotlin.test.assertEquals
